+++ playground/src/app.js
@@ -26,7 +26,7 @@ import * as AutoImporter from './auto-importer/auto-importer.js';
 import * as ComponentsProvider from './components-provider/components-provider.js';
 import * as EmailLoader from './email-loader/email-loader.js';
 import * as CspHashCalculator from './csp-hash-calculator/csp-hash-calculator.js';
-import * as ErrorList from './error-list/error-list.js';
+import * as ErrorListFlyIn from './error-list/error-list-fly-in.js';
 import * as StateView from './state-view/state-view.js';
 import * as ImportURL from './import-url/import-url.js';
 import * as Validator from './validator/validator.js';
@@ -74,9 +74,9 @@ StateView.createStateView(stateListContainer, stateIndicator);
 // configure error list behavior
 const errorIndicator = document.getElementById('error-indicator');
 const errorListContainer = document.getElementById('error-list');
-ErrorList.createErrorList(errorListContainer, errorIndicator);
+ErrorListFlyIn.createErrorList(errorListContainer, errorIndicator);
 
-events.subscribe(ErrorList.EVENT_ERROR_SELECTED, (error) =>
+events.subscribe(ErrorListFlyIn.EVENT_ERROR_SELECTED, (error) =>
   editor.setCursorAndFocus(error.line, error.col)
 );
